requirements.txt
streamlit
python-docx
pdfplumber
python-dotenv
google-generativeai